requirements.txt
requests
python-dotenv
PIL
numpy